Moved RetroFit interface to a more appropriate package.
